pyproject.toml
[project]
name = "rbyte"
version = "0.7.0"
description = "Multimodal PyTorch dataset library"
authors = [{ name = "Evgenii Gorchakov", email = "evgenii@yaak.ai" }]
maintainers = [{ name = "Evgenii Gorchakov", email = "evgenii@yaak.ai" }]
dependencies = [
  "tensordict>=0.6.2",
  "torch",
  "numpy",
  "polars>=1.14.0",
  "pydantic>=2.9.2",
  "more-itertools>=10.5.0",
  "hydra-core>=1.3.2",
  "optree>=0.13.1",
  "cachetools>=5.5.0",
  "diskcache>=5.6.3",
  "jaxtyping>=0.2.34",
  "parse>=1.20.2",
  "structlog>=24.4.0",
  "xxhash>=3.5.0",
  "tqdm>=4.66.5",
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "Apache-2.0" }
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
repo = "https://github.com/yaak-ai/rbyte"

[project.optional-dependencies]
build = ["hatchling>=1.25.0", "grpcio-tools>=1.62.0", "protoletariat==3.2.19"]
visualize = ["rerun-sdk[notebook]>=0.20.0"]
mcap = [
  "mcap>=1.2.1",
  "mcap-ros2-support>=0.5.5",
  "protobuf",
  "mcap-protobuf-support>=0.5.2",
]
yaak = ["protobuf", "ptars>=0.0.3"]
jpeg = ["simplejpeg>=1.7.6"]
video = [
  "python-vali>=4.2.0.post0; sys_platform == 'linux'",
  "video-reader-rs>=0.2.1",
]
hdf5 = ["h5py>=3.12.1"]
rrd = ["rerun-sdk>=0.20.0", "pyarrow-stubs"]

[project.scripts]
rbyte-visualize = 'rbyte.scripts.visualize:main'

[build-system]
requires = [
  "hatchling>=1.25.0",
  "grpcio-tools>=1.62.0",
  "protoletariat==3.2.19",
]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "wat-inspector>=0.4.2",
  "lovely-tensors>=0.1.17",
  "pudb>=2024.1.2",
  "ipython>=8.29.0",
  "ipython-autoimport>=0.5",
  "pytest>=8.3.3",
  "testbook>=0.4.2",
  "ipykernel>=6.29.5",
]

[tool.uv.sources]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
reproducible = true

[tool.hatch.build.targets.sdist]
include = ["src/rbyte"]
exclude = ["src/rbyte/io/table/yaak/idl-repo"]
artifacts = ["src/rbyte/io/table/yaak/proto/*_pb2.py*"]

[tool.hatch.build.targets.sdist.hooks.custom]
enable-by-default = true
require-runtime-features = ["build"]

[tool.hatch.build.targets.wheel]
packages = ["src/rbyte"]
artifacts = ["src/rbyte/io/table/yaak/proto/*_pb2.py*"]

[tool.basedpyright]
typeCheckingMode = "all"
enableTypeIgnoreComments = true
reportMissingTypeStubs = "none"
reportAny = "none"
reportIgnoreCommentWithoutRule = "error"

venvPath = "."
venv = ".venv"
include = ["src/"]
ignore = ["src/**/*_pb2.py", "**/*.pyi"]

[tool.ruff]
extend-exclude = ['*_pb2.py']
output-format = "grouped"
preview = true
fix = true
unsafe-fixes = true
target-version = "py312"

[tool.ruff.format]
preview = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
preview = true
select = ["ALL"]
fixable = ["ALL"]
ignore = ["D", "CPY", "COM812", "F722", "PD901", "ISC001", "TD"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.pytest.ini_options]
addopts = ["-ra", "-vv", "--import-mode=importlib"]
testpaths = ["tests"]
